Refactor folder structure (#2100)
