+++ apps/browser-extension/src/entrypoints/popup/components/Items/ItemFilterDropdown.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 
 import { ITEM_TYPE_OPTIONS } from '@/entrypoints/popup/components/Items/ItemTypeSelector';
 
-import { isItemTypeFilter, type ItemFilterType } from '@/utils/itemFilters';
+import { isItemTypeFilter, type ItemFilterType } from '@/utils/ItemFilters';
 
 /**
  * Filter selection that includes the dedicated "deleted" page as an option too.
