+++ apps/browser-extension/src/utils/FolderUtils.ts
