+++ apps/browser-extension/src/utils/PasswordLengthSlider.ts
